Polkakernel fixes for CoreVM.
- Make all methods accept `&mut self`.
- Add `Machine::touch_memory`.
- Move `Kernel::init` to `Machine::init`.
- Remove `start` argument from `Machine::init`.
